refactor: split compile into Context's and lib.rs
