+++ src/context.rs
@@ -1,22 +1,10 @@
 #![allow(dead_code)]
 use llvm_sys::{
-    core::{
-        LLVMContextCreate, LLVMContextDispose, LLVMDisposeMessage, LLVMDisposeModule,
-        LLVMPrintModuleToFile,
-    },
-    error::LLVMGetErrorMessage,
-    target::{
-        LLVM_InitializeAllAsmPrinters, LLVM_InitializeAllTargetInfos, LLVM_InitializeAllTargetMCs,
-        LLVM_InitializeAllTargets,
-    },
+    core::LLVMDisposeMessage,
     target_machine::{
-        LLVMCodeGenFileType, LLVMCodeGenOptLevel, LLVMCodeModel, LLVMCreateTargetMachine,
-        LLVMDisposeTargetMachine, LLVMGetDefaultTargetTriple, LLVMGetHostCPUFeatures,
-        LLVMGetHostCPUName, LLVMGetTargetFromTriple, LLVMRelocMode, LLVMTargetMachineEmitToFile,
+        LLVMCodeGenOptLevel, LLVMCodeModel, LLVMCreateTargetMachine, LLVMGetDefaultTargetTriple,
+        LLVMGetHostCPUFeatures, LLVMGetHostCPUName, LLVMGetTargetFromTriple, LLVMRelocMode,
         LLVMTargetRef,
-    },
-    transforms::pass_builder::{
-        LLVMCreatePassBuilderOptions, LLVMDisposePassBuilderOptions, LLVMRunPasses,
     },
 };
 use melior::{
@@ -34,13 +22,11 @@ use melior::{
     utility::{register_all_dialects, register_all_llvm_translations, register_all_passes},
     Context as MeliorContext,
 };
-use mlir_sys::mlirTranslateModuleToLLVMIR;
 use std::{
-    ffi::{CStr, CString},
+    ffi::CStr,
     mem::MaybeUninit,
     path::PathBuf,
     ptr::{addr_of_mut, null_mut},
-    sync::OnceLock,
 };
 
 use crate::{
@@ -71,15 +57,7 @@ impl Context {
         Self { melior_context }
     }
 
-    pub fn compile(&self, program: Vec<Operation>) -> Result<MLIRModule, String> {
-        static INITIALIZED: OnceLock<()> = OnceLock::new();
-        INITIALIZED.get_or_init(|| unsafe {
-            LLVM_InitializeAllTargets();
-            LLVM_InitializeAllTargetInfos();
-            LLVM_InitializeAllTargetMCs();
-            LLVM_InitializeAllAsmPrinters();
-        });
-
+    pub fn compile(&self, program: &[Operation]) -> Result<MLIRModule, String> {
         let target_triple = get_target_triple();
 
         let context = &self.melior_context;
@@ -207,143 +185,6 @@ pub fn get_data_layout_rep() -> Result<String, String> {
         let data_layout_str =
             CStr::from_ptr(llvm_sys::target::LLVMCopyStringRepOfTargetData(data_layout));
         Ok(data_layout_str.to_string_lossy().into_owned())
-    }
-}
-
-/// Converts a module to an object.
-/// The object will be written to the specified target path.
-/// TODO: error handling
-///
-/// Returns the path to the object.
-// TODO: pass options to the function
-pub fn compile_to_object(module: &MLIRModule<'_>) -> Result<PathBuf, String> {
-    // TODO: put a proper target_file here
-    let target_file = PathBuf::from("output.o");
-    // let target_file = session.output_file.with_extension("o");
-
-    // TODO: Rework so you can specify target and host features, etc.
-    // Right now it compiles for the native cpu feature set and arch
-    unsafe {
-        let llvm_context = LLVMContextCreate();
-
-        let op = module.melior_module.as_operation().to_raw();
-
-        let llvm_module = mlirTranslateModuleToLLVMIR(op, llvm_context as *mut _) as *mut _;
-
-        let mut null = null_mut();
-        let mut error_buffer = addr_of_mut!(null);
-
-        let target_triple = LLVMGetDefaultTargetTriple();
-
-        let target_cpu = LLVMGetHostCPUName();
-
-        let target_cpu_features = LLVMGetHostCPUFeatures();
-
-        let mut target: MaybeUninit<LLVMTargetRef> = MaybeUninit::uninit();
-
-        if LLVMGetTargetFromTriple(target_triple, target.as_mut_ptr(), error_buffer) != 0 {
-            let error = CStr::from_ptr(*error_buffer);
-            let err = error.to_string_lossy().to_string();
-            LLVMDisposeMessage(*error_buffer);
-            Err(err)?;
-        } else if !(*error_buffer).is_null() {
-            LLVMDisposeMessage(*error_buffer);
-            error_buffer = addr_of_mut!(null);
-        }
-
-        let target = target.assume_init();
-
-        let machine = LLVMCreateTargetMachine(
-            target,
-            target_triple.cast(),
-            target_cpu.cast(),
-            target_cpu_features.cast(),
-            LLVMCodeGenOptLevel::LLVMCodeGenLevelNone,
-            LLVMRelocMode::LLVMRelocDefault,
-            LLVMCodeModel::LLVMCodeModelDefault,
-        );
-
-        let opts = LLVMCreatePassBuilderOptions();
-        let opt = 0;
-        let passes = CString::new(format!("default<O{opt}>")).unwrap();
-        let error = LLVMRunPasses(llvm_module as *mut _, passes.as_ptr(), machine, opts);
-        if !error.is_null() {
-            let msg = LLVMGetErrorMessage(error);
-            let msg = CStr::from_ptr(msg);
-            Err(msg.to_string_lossy().into_owned())?;
-        }
-
-        LLVMDisposePassBuilderOptions(opts);
-
-        // Output the LLVM IR
-        let filename = CString::new(
-            target_file
-                .with_extension("ll")
-                .as_os_str()
-                .to_string_lossy()
-                .as_bytes(),
-        )
-        .unwrap();
-        if LLVMPrintModuleToFile(llvm_module, filename.as_ptr(), error_buffer) != 0 {
-            let error = CStr::from_ptr(*error_buffer);
-            let err = error.to_string_lossy().to_string();
-            LLVMDisposeMessage(*error_buffer);
-            Err(err)?;
-        } else if !(*error_buffer).is_null() {
-            LLVMDisposeMessage(*error_buffer);
-            error_buffer = addr_of_mut!(null);
-        }
-
-        // Output the object file
-        let filename = CString::new(target_file.as_os_str().to_string_lossy().as_bytes()).unwrap();
-        let ok = LLVMTargetMachineEmitToFile(
-            machine,
-            llvm_module,
-            filename.as_ptr().cast_mut(),
-            LLVMCodeGenFileType::LLVMObjectFile, // object (binary) or assembly (textual)
-            error_buffer,
-        );
-
-        if ok != 0 {
-            let error = CStr::from_ptr(*error_buffer);
-            let err = error.to_string_lossy().to_string();
-            LLVMDisposeMessage(*error_buffer);
-            Err(err)?;
-        } else if !(*error_buffer).is_null() {
-            LLVMDisposeMessage(*error_buffer);
-        }
-
-        // Output the assembly
-        let filename = CString::new(
-            target_file
-                .with_extension("asm")
-                .as_os_str()
-                .to_string_lossy()
-                .as_bytes(),
-        )
-        .unwrap();
-        let ok = LLVMTargetMachineEmitToFile(
-            machine,
-            llvm_module,
-            filename.as_ptr().cast_mut(),
-            LLVMCodeGenFileType::LLVMAssemblyFile,
-            error_buffer,
-        );
-
-        if ok != 0 {
-            let error = CStr::from_ptr(*error_buffer);
-            let err = error.to_string_lossy().to_string();
-            LLVMDisposeMessage(*error_buffer);
-            Err(err)?;
-        } else if !(*error_buffer).is_null() {
-            LLVMDisposeMessage(*error_buffer);
-        }
-
-        LLVMDisposeTargetMachine(machine);
-        LLVMDisposeModule(llvm_module);
-        LLVMContextDispose(llvm_context);
-
-        Ok(target_file)
     }
 }
