+++ src/main.rs
@@ -1,8 +1,4 @@
-use evm_mlir::{
-    context::{compile_to_object, Context},
-    opcodes::Operation,
-};
-use std::path::PathBuf;
+use evm_mlir::{compile, compile_to_object, context::Context, link_binary, opcodes::Operation};
 
 fn main() {
     let args: Vec<String> = std::env::args().collect();
@@ -10,34 +6,11 @@ fn main() {
     let bytecode = std::fs::read(path).expect("Could not read file");
     let operations = Operation::from_bytecode(bytecode);
 
-    let context = Context::new();
-
-    println!("Creating MLIR module");
-
-    let result = context.compile(operations).unwrap();
-
-    println!("Compiling with LLVM");
-
-    let object_file = compile_to_object(&result).unwrap();
+    let object_file = compile(operations);
 
     println!("Linking...");
 
     link_binary(object_file);
 
     println!("Done!");
-}
-
-fn link_binary(object_file: PathBuf) {
-    let args = vec![
-        "-L/usr/local/lib",
-        "-L/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/lib",
-        object_file.to_str().unwrap(),
-        "-o",
-        "output",
-        "-lSystem",
-    ];
-    let mut linker = std::process::Command::new("ld");
-    let proc = linker.args(args).spawn().unwrap();
-    let output = proc.wait_with_output().unwrap();
-    assert!(output.status.success());
 }
